fix: formatting logic moved from view to viewModel
